Cleaned up some PATH sludge
Built a local version of Go 1.4, since it doesn't seem likely it will
ever appear in the package manager.

This made me aware of some ugliness in the shell path that need some
attention.

The distro zsh config files add everything in ~/bin to your path now. At
least here in Red Hat land. That was creating a lot of duplication.

It also appears that the PATH+= syntax is not dependable. Nearly threw
the laptop through the wall figuring that out ...