Mutt: do not use Mail-Followup-To header
This is a non-standard feature that is on by default. It will inject a
separate header into any message you send to a subscribed list. The idea
was to prevent separate replies from lists, but it effectively drops me
from the CC on any list and I end up missing mail.

Plus it annoys other people on some lists I'm subscribed to.