Tell Neovim when editor focus is regained
The 'autoread' setting is enabled by default in Neovim. But Neovim runs as a
terminal application or as a client embedded in a GUI; it is not aware
of when application focus changes. To get the equivalent GVim behavior
so that 'autoread' works we have to give nvim a hint that focus has
been regained.